Repair merge problem from rebase
